Rework UpdateService creation to avoid calling Toggl.UpdateChannel() before DB is initialized (win)
